Add a new fixture scope 'session_prerun'.
It can be used for code that needs to be run before the test session is started: an exception
in a fixture with this scope will prevent the tests to be executed.
Add a new UserError exception: if this exception is raised from session_prerun scoped fixture
or a project before/after_run hook then only the string representation of the exception will be
display to the user.
Remove the --show-stacktrace option of lcc run: the full stacktrace of custom code will be
systematically displayed if the class exception is different from UserError.